moves motion border util to general place
