+++ src/ophys_etl/utils/motion_border.py
